// frontend/sdk/models/GauntletHistorySummaryDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Unpaused Spotify API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface GauntletHistorySummaryDto
 */
export interface GauntletHistorySummaryDto {
    /**
     * Total runs in the selected scope
     * @type {number}
     * @memberof GauntletHistorySummaryDto
     */
    totalRuns: number;
    /**
     * Best score in the selected scope
     * @type {number}
     * @memberof GauntletHistorySummaryDto
     */
    bestScore: number;
    /**
     * Average score in the selected scope
     * @type {number}
     * @memberof GauntletHistorySummaryDto
     */
    averageScore: number;
    /**
     * Total correct answers across all runs in scope
     * @type {number}
     * @memberof GauntletHistorySummaryDto
     */
    totalCorrectAnswers: number;
}

/**
 * Check if a given object implements the GauntletHistorySummaryDto interface.
 */
export function instanceOfGauntletHistorySummaryDto(value: object): value is GauntletHistorySummaryDto {
    if (!('totalRuns' in value) || value['totalRuns'] === undefined) return false;
    if (!('bestScore' in value) || value['bestScore'] === undefined) return false;
    if (!('averageScore' in value) || value['averageScore'] === undefined) return false;
    if (!('totalCorrectAnswers' in value) || value['totalCorrectAnswers'] === undefined) return false;
    return true;
}

export function GauntletHistorySummaryDtoFromJSON(json: any): GauntletHistorySummaryDto {
    return GauntletHistorySummaryDtoFromJSONTyped(json, false);
}

export function GauntletHistorySummaryDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): GauntletHistorySummaryDto {
    if (json == null) {
        return json;
    }
    return {
        
        'totalRuns': json['totalRuns'],
        'bestScore': json['bestScore'],
        'averageScore': json['averageScore'],
        'totalCorrectAnswers': json['totalCorrectAnswers'],
    };
}

export function GauntletHistorySummaryDtoToJSON(json: any): GauntletHistorySummaryDto {
    return GauntletHistorySummaryDtoToJSONTyped(json, false);
}

export function GauntletHistorySummaryDtoToJSONTyped(value?: GauntletHistorySummaryDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'totalRuns': value['totalRuns'],
        'bestScore': value['bestScore'],
        'averageScore': value['averageScore'],
        'totalCorrectAnswers': value['totalCorrectAnswers'],
    };
}